requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
apturl==0.5.2
async-timeout==4.0.3
attrs==19.3.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2024.8.30
chardet==3.0.4
click==8.1.7
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.14.0
httpx==0.27.2
idna==2.8
importlib-metadata==1.5.0
Jinja2==3.1.4
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
magic-filter==1.0.12
MarkupSafe==2.1.5
more-itertools==4.2.0
multidict==6.0.5
netifaces==0.10.4
nodeenv==1.9.1
oauthlib==3.1.0
olefile==0.46
pexpect==4.6.0
Pillow==7.0.0
prisma==0.15.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
pydantic==2.8.2
pydantic_core==2.20.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
StrEnum==0.4.15
systemd-python==234
texttable==1.6.2
tomlkit==0.13.2
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vboxapi==1.0
wadllib==1.3.3
websocket-client==0.53.0
xkit==0.0.0
yarl==1.9.8
zipp==1.0.0
